Refactor mempool tx generation code into its own function

+++ code/crates/actors/src/util/value_builder.rs
@@ -53,7 +53,7 @@ pub mod test {
     pub struct TestParams {
         pub max_block_size: ByteSize,
         pub tx_size: ByteSize,
-        pub txs_per_part: u64,
+        pub txs_per_part: usize,
         pub time_allowance_factor: f32,
         pub exec_time_per_tx: Duration,
     }
